+++ download.go
@@ -11,7 +11,7 @@ import (
 	"sync"
 
 	alpm "github.com/Jguer/go-alpm"
-	"github.com/Jguer/yay/v9/pkg/types"
+	"github.com/Jguer/yay/v9/pkg/multierror"
 )
 
 const gitDiffRefName = "AUR_SEEN"
@@ -265,7 +265,7 @@ func getPkgbuilds(pkgs []string) error {
 func getPkgbuildsfromABS(pkgs []string, path string) (bool, error) {
 	var wg sync.WaitGroup
 	var mux sync.Mutex
-	var errs types.MultiError
+	var errs multierror.MultiError
 	names := make(map[string]string)
 	missing := make([]string, 0)
 	downloaded := 0
